[Refactor] #94 LIVD-56 - ConcertInfoCarousel 분리 및 레이아웃 개선
- ConcertInfoCarousel을 별도 파일로 분리
- 탭/스와이프 제스처 분리 (onTapGesture + simultaneousGesture)
- LazyVStack 내 이미지 오버플로우 수정 (Color.clear 컨테이너 방식)
- merchandiseSection 패딩 구조 정리
